Rename service to vm_service
