+++ crates/devtools/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod io_extension;
 pub mod protocols;
-pub mod service;
 pub(crate) mod util;
+pub mod vm_service;
