+++ crates/devtools/src/vm_service.rs
